+++ __docs__/components/gallery/tiles/link-tile.tsx
@@ -17,8 +17,7 @@ export default function LinkTile(props: CommonTileProps) {
             {...props}
         >
             <View style={styles.centerContent}>
-                {/* eslint-disable-next-line jsx-a11y/anchor-is-valid -- TODO: Address a11y error */}
-                <Link href="#">This is a link</Link>
+                <Link href="#link">This is a link</Link>
             </View>
         </ComponentTile>
     );