+++ __docs__/wonder-blocks-modal/modal-header.stories.tsx
@@ -273,12 +273,10 @@ export const WithBreadcrumbs: StoryComponentType = {
                         breadcrumbs={
                             <Breadcrumbs>
                                 <BreadcrumbsItem>
-                                    {/* eslint-disable-next-line jsx-a11y/anchor-is-valid -- TODO: Address a11y error */}
-                                    <Link href="">Course</Link>
+                                    <Link href="#course">Course</Link>
                                 </BreadcrumbsItem>
                                 <BreadcrumbsItem>
-                                    {/* eslint-disable-next-line jsx-a11y/anchor-is-valid -- TODO: Address a11y error */}
-                                    <Link href="">Unit</Link>
+                                    <Link href="#unit">Unit</Link>
                                 </BreadcrumbsItem>
                                 <BreadcrumbsItem>Lesson</BreadcrumbsItem>
                             </Breadcrumbs>