+++ __docs__/wonder-blocks-popover/popover-content.stories.tsx
@@ -53,6 +53,9 @@ const styles = StyleSheet.create({
 
 type StoryComponentType = StoryObj<typeof PopoverContent>;
 
+/**
+ * Default popover variant that displays text-only.
+ */
 export const Default: StoryComponentType = {
     args: {
         title: "A simple popover",
@@ -64,14 +67,11 @@ export const Default: StoryComponentType = {
 
 Default.storyName = "Default (text)";
 
-Default.parameters = {
-    docs: {
-        description: {
-            story: "Default popover variant that displays text-only.",
-        },
-    },
-};
-
+/**
+ * Text-only variant with added emphasis.
+ * **NOTE:** When using this variant, make sure to apply the `light`
+ * prop to each button
+ */
 export const Emphasized: StoryComponentType = {
     args: {
         title: "Popover with emphasis",
@@ -93,17 +93,15 @@ export const Emphasized: StoryComponentType = {
     render: (args) => <PopoverContent {...args} />,
 };
 
-Emphasized.parameters = {
-    docs: {
-        description: {
-            story:
-                `Text-only variant with added emphasis.\n\n` +
-                `**NOTE:** When using this variant, make sure to apply the \`light\`
-        prop to each button`,
-        },
-    },
-};
-
+/**
+ * Decorate the popover with an illustrated icon. You need to pass an `icon`
+ * prop with the following constraints:
+ * - string: The URL of the icon asset
+ * - `<img>` or `<svg>`: Make sure to define a width
+ *
+ * When passing in a url for the `icon` prop, use the `iconAlt` prop to provide
+ * alternative text for the icon if it communicates meaning.
+ */
 export const WithIcon: StoryComponentType = {
     args: {
         title: "Popover with Icon",
@@ -113,14 +111,9 @@ export const WithIcon: StoryComponentType = {
     render: (args) => <PopoverContent {...args} />,
 };
 
-WithIcon.parameters = {
-    docs: {
-        description: {
-            story: "Decorate the popover with an illustrated icon. You need to pass an `icon` prop with the following constraints:\n\n - string: The URL of the icon asset\n - `<img>` or `<svg>`: Make sure to define a width",
-        },
-    },
-};
-
+/**
+ * Call attention to the popover using a full-bleed illustration.
+ */
 export const WithIllustration: StoryComponentType = {
     args: {
         title: "Popover with Illustration",
@@ -137,12 +130,4 @@ export const WithIllustration: StoryComponentType = {
         closeButtonVisible: true,
     },
     render: (args) => <PopoverContent {...args} />,
-};
-
-WithIllustration.parameters = {
-    docs: {
-        description: {
-            story: "Call attention to the popover using a full-bleed illustration.",
-        },
-    },
 };