+++ packages/wonder-blocks-form/src/components/text-area.tsx
@@ -187,6 +187,17 @@ type TextAreaProps = AriaProps & {
 
 const StyledTextarea = addStyle("textarea");
 
+/**
+ * A TextArea is an element used to accept text from the user.
+ *
+ * Make sure to provide a label for the field. This can be done by either:
+ * - (recommended) Using the **LabeledField** component to provide a label,
+ * description, and/or error message for the field
+ * - Using a `label` html tag with the `htmlFor` prop set to the unique id of
+ * the field
+ * - Using an `aria-label` attribute on the field
+ * - Using an `aria-labelledby` attribute on the field
+ */
 const TextArea = React.forwardRef<HTMLTextAreaElement, TextAreaProps>(
     function TextArea(
         props: TextAreaProps,
@@ -313,8 +324,10 @@ const states = {
         foreground: semanticColor.text.primary,
     },
     disabled: {
-        border: semanticColor.border.primary,
-        background: semanticColor.action.disabled.secondary,
+        border: semanticColor.action.secondary.disabled.border,
+        background: semanticColor.action.secondary.disabled.background,
+        // NOTE: This color is specific for form fields.
+        // TODO(WB-1895): Revisit disabled styles.
         foreground: semanticColor.text.secondary,
     },
     // Form validation error state
@@ -347,8 +360,8 @@ const styles = StyleSheet.create({
     },
     defaultFocus: {
         ":focus-visible": {
-            borderColor: semanticColor.border.focus,
-            outline: `${border.width.hairline}px solid ${semanticColor.border.focus}`,
+            borderColor: semanticColor.focus.outer,
+            outline: `${border.width.hairline}px solid ${semanticColor.focus.outer}`,
             // Negative outline offset so it focus outline is not cropped off if
             // an ancestor element has overflow: hidden
             outlineOffset: -2,
@@ -363,8 +376,7 @@ const styles = StyleSheet.create({
         },
         cursor: "not-allowed",
         ":focus-visible": {
-            // TODO(WB-1856): Verify if we can use the global focus color
-            outline: `${border.width.thin}px solid ${semanticColor.action.disabled.default}`,
+            outline: `${border.width.thin}px solid ${semanticColor.focus.outer}`,
             outlineOffset: -3,
         },
     },
@@ -376,8 +388,7 @@ const styles = StyleSheet.create({
             color: semanticColor.text.secondary,
         },
         ":focus-visible": {
-            // TODO(WB-1856): Verify if we can use the global focus color
-            outlineColor: states.error.border,
+            outline: `${border.width.thin}px solid ${semanticColor.focus.outer}`,
             borderColor: states.error.border,
         },
     },